requirements.txt
accelerate==1.10.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.0
aiosignal==1.4.0
annotated-types==0.7.0
anthropic==0.69.0
anyio==4.10.0
asyncpg==0.30.0
attrs==25.4.0
beautifulsoup4==4.14.2
cachetools==6.2.1
certifi==2025.8.3
charset-normalizer==3.4.3
colorama==0.4.6
dataclasses-json==0.6.7
distro==1.9.0
docstring_parser==0.17.0
dotenv==0.9.9
filelock==3.13.1
filetype==1.2.0
frozenlist==1.8.0
fsspec==2024.6.1
google-ai-generativelanguage==0.7.0
# google-generativeai==0.8.3
google-api-core==2.26.0
google-api-python-client==2.184.0
google-auth==2.41.1
google-auth-httplib2==0.2.0
googleapis-common-protos==1.70.0
greenlet==3.2.4
grpcio==1.75.1
grpcio-status==1.75.1
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.35.3
idna==3.10
Jinja2==3.1.4
jiter==0.10.0
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.27
langchain-anthropic==0.3.22
langchain-community==0.3.31
langchain-core==0.3.79
langchain-experimental==0.3.4
langchain-google-genai==2.1.12
langchain-huggingface==0.3.1
langchain-openai==0.3.32
langchain-postgres==0.0.15
langchain-text-splitters==0.3.11
langsmith==0.4.26
MarkupSafe==2.1.5
marshmallow==3.26.1
mpmath==1.3.0
multidict==6.7.0
mypy_extensions==1.1.0
networkx==3.3
numpy==1.26.4
openai==1.106.1
orjson==3.11.3
packaging==25.0
pgvector==0.3.6
pillow==11.0.0
propcache==0.4.1
proto-plus==1.26.1
protobuf==6.32.1
psutil==7.1.0
psycopg==3.2.10
psycopg-binary==3.2.10
psycopg-pool==3.2.6
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic-settings==2.11.0
pydantic_core==2.33.2
pyparsing==3.2.5
pypdf==6.1.1
python-dotenv==1.1.1
PyYAML==6.0.2
rank-bm25==0.2.2
regex==2025.9.1
requests==2.32.5
requests-toolbelt==1.0.0
rsa==4.9.1
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.2
sentence-transformers==5.1.1
setuptools==70.2.0
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.43
sympy==1.13.1
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.11.0
tokenizers==0.19.1
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tqdm==4.67.1
transformers==4.42.4
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
yarl==1.22.0
zstandard==0.24.0
langgraph==0.2.68
fastapi==0.115.6
uvicorn[standard]==0.34.0

